asr/Interaction/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.2
fsspec==2025.7.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysql-connector-python==9.0.0
numpy==2.2.6
openpyxl==3.1.2
packaging==25.0
pandas==2.3.0
phone==0.4.5
pika==1.3.2
Pygments==2.19.1
pypinyin==0.54.0
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==3.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
